Moved the exception
